moved HyetoMethod Enum from 'hyetograph.py' to 'enum_rain_input_type.py'
